% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pheatmap_df.R
\name{make_pheatmap_df}
\alias{make_pheatmap_df}
\title{A function to make a data frame that will be in the library pheatmap's color limits that you want.}
\usage{
make_pheatmap_df(x, min = -3, max = 3)
}
\arguments{
\item{x}{the pheatmap data matrix input}

\item{min}{the minimum value in your color scheme}

\item{max}{the maximum value in your color scheme}
}
\description{
This function allows you to write out tsv files with default options that make sense..
}
\examples{
make_pheatmap_df()
}
\keyword{pheatmap}
\keyword{table}

