% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pheatmap_groups.R
\name{make_pheatmap_groups}
\alias{make_pheatmap_groups}
\title{Annotate your group clusters easier.}
\usage{
make_pheatmap_groups(hclust_object, numClusters = 4)
}
\arguments{
\item{hclust_object}{the output from a call to hclust}

\item{numClusters}{the number of clusters you want to have assigned to a group}
}
\description{
This function creates the data frame for input into pheatmaps 'annotation_row' or 'annotation_col' option
}
\examples{
make_pheatmap_groups()
}
\keyword{groups}
\keyword{pheatmap}

