% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyFilePath.R
\name{prettyFilePath}
\alias{prettyFilePath}
\title{Prettyfies file paths from '?list.files' output}
\usage{
prettyFilePath(x, prepend = "", sep = "\\n\\n")
}
\arguments{
\item{x}{output from '?list.files'}

\item{prepend}{anything you would like to prepend to your filepath i.e. however you link out}

\item{sep}{what characters are used to separate the output from '?list.files'}
}
\description{
Makes file paths pretty to include in reports as links, or to send to people.
}
\examples{
pretty_file_path(list.files('.', full.names = TRUE))
}
\keyword{file}
\keyword{path}
\keyword{pretty}

