% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyLinkFile.R
\name{prettyLinkFile}
\alias{prettyLinkFile}
\title{Makes pretty links}
\usage{
prettyLinkFile(x, y, sep = "\\n\\n")
}
\arguments{
\item{x}{output from '?list.files' that you want to be the text for the link}

\item{y}{file path to the location you want to link to}

\item{sep}{what the separator for '?cat' should be}
}
\description{
Automates making mass links to output files. Works well with '?mapply', see example.
}
\examples{
invisible(mapply(prettyFileLink, 
x = list.files('.'), 
y = list.files('.', full.names = TRUE)))
}
\keyword{file}
\keyword{path}
\keyword{pretty}

