% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\title{Calculate Standard Error}
\usage{
se(x)
}
\arguments{
\item{x}{a vector to calculate the error for.}
}
\description{
This function calculates the standard error.
}
\examples{
vec <- seq(1, 50, by = 3)
se(vec)
}
\keyword{error}
\keyword{standard}

