% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitstr2vec.R
\name{splitstr2vec}
\alias{splitstr2vec}
\title{Turn ?strsplit lists into vectors}
\usage{
splitstr2vec(x, split = " ")
}
\arguments{
\item{x}{a character vector to split}

\item{split}{input to ?strsplit}
}
\description{
A wrapper for ?strsplit that will split the your character vector and only return the first element of the resultant list.
Useful for breaking up a character vector when you only want output type, instead of having the vector embedded in a list.
}
\examples{
test <- 'Hello there world! I like cats and unicorns.'
strsplit2vec(test)
strsplit2vec(test)[2] 
}
\keyword{strsplit}
\keyword{vector}

