% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteTable.R
\name{WriteTable}
\alias{WriteTable}
\title{Another Table Writing Function}
\usage{
WriteTable(x, file)
}
\arguments{
\item{x}{the data frame that you want to write out.}

\item{file}{the path of the file you are writing to}
}
\description{
This function allows you to write out tsv files with default options that make sense..
}
\examples{
WriteTable()
}
\keyword{table}
\keyword{write}

